Changed from fixed4 to half4

+++ Resources/Shaders/NormalMapDecoder.shader
@@ -41,7 +41,7 @@
 
 			fixed4 frag (v2f i) : SV_Target
 			{
-				fixed4 col = tex2D(_MainTex, i.uv);
+				half4 col = tex2D(_MainTex, i.uv);
 
 				col.xyz = (UnpackNormal(col) + 1) * 0.5;
 				col.w = 1;
